+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - New logo added
 - Additional unit tests added
+- GgpkRecords.From(Stream) method added
 
 ## 0.1.0 (2018-12-29)